Split code and continue refactor
